+++ playwright.config.ts
@@ -1,4 +1,8 @@
 import { defineConfig, devices } from "@playwright/test";
+import dotenv from "dotenv";
+
+dotenv.config({ path: ".env.test" });
+dotenv.config({ path: ".env.local" });
 
 export default defineConfig({
   testDir: "./tests",